+++ pyriemann/utils/mean.py
@@ -275,14 +275,17 @@ def mean_kullback_sym(X=None, sample_weight=None, covmats=None):
     return M
 
 
-def mean_logcholesky(X, sample_weight=None):
+def mean_logchol(X, sample_weight=None):
     r"""Mean of SPD/HPD matrices according to the log-Cholesky metric.
 
-    Log-Cholesky mean is [1]_:
+    Log-Cholesky mean :math:`\mathbf{M}` is
+    :math:`\mathbf{M} = \mathbf{L} \mathbf{L}^H`,
+    where :math:`\mathbf{L}` is computed as [1]_:
 
     .. math::
-        \mathbf{M} =\sum_i w_i \text{lower}(\text{chol}(\mathbf{X}_i})) +
-        \exp{\sum_i w_i \log{\text{diag}\text{chol}({\mathbf{X}_i}}) }
+        \mathbf{L} = \sum_i w_i \text{lower}(\text{chol}(\mathbf{X}_i)) +
+        \exp \left( \sum_i w_i \log(\text{diag}(\text{chol}(\mathbf{X}_i)))
+        \right)
 
     Parameters
     ----------
@@ -316,17 +319,23 @@ def mean_logcholesky(X, sample_weight=None):
 
     X_chol = np.linalg.cholesky(X)
 
-    tr0, tr1 = np.tril_indices(n_channels, -1)
-    diag0, diag1 = np.diag_indices(n_channels)
-
     mean = np.zeros(X.shape[-2:], dtype=X.dtype)
 
-    mean[..., tr0, tr1] = np.average(X_chol[..., tr0, tr1], axis=0,
-                                     weights=sample_weight)
-    mean[diag0, diag1] = np.exp(
-        np.average(np.log(X_chol[..., diag0, diag1]),
-                   axis=0, weights=sample_weight))
-    return mean @ mean.conj().swapaxes(-2, -1)
+    tri0, tri1 = np.tril_indices(n_channels, -1)
+    mean[tri0, tri1] = np.average(
+        X_chol[:, tri0, tri1],
+        axis=0,
+        weights=sample_weight,
+    )
+
+    diag0, diag1 = np.diag_indices(n_channels)
+    mean[diag0, diag1] = np.exp(np.average(
+        np.log(X_chol[:, diag0, diag1]),
+        axis=0,
+        weights=sample_weight,
+    ))
+
+    return mean @ mean.conj().T
 
 
 def mean_logdet(X=None, tol=10e-5, maxiter=50, init=None, sample_weight=None,
@@ -725,7 +734,7 @@ mean_functions = {
     "identity": mean_identity,
     "kullback_sym": mean_kullback_sym,
     "logdet": mean_logdet,
-    "logcholesky": mean_logcholesky,
+    "logchol": mean_logchol,
     "logeuclid": mean_logeuclid,
     "riemann": mean_riemann,
     "wasserstein": mean_wasserstein,
@@ -744,8 +753,9 @@ def mean_covariance(X=None, metric="riemann", sample_weight=None, covmats=None,
         Set of matrices.
     metric : string | callable, default="riemann"
         Metric for mean estimation, can be: "ale", "alm", "euclid", "harmonic",
-        "identity", "kullback_sym", "logdet", "logeuclid", "riemann",
-        "wasserstein", or a callable function.
+        "identity", "kullback_sym", "logchol", "logdet", "logeuclid",
+        "riemann", "wasserstein",
+        or a callable function.
     sample_weight : None | ndarray, shape (n_matrices,), default=None
         Weights for each matrix. If None, it uses equal weights.
     **kwargs : dict